+++ lib/main.ts
@@ -1,7 +1,8 @@
 // collisions
-export * from './core/collisions-2D';
+export * from './core/collisions';
 
 // shapes
 export { Circle } from './core/shapes/circle';
+export { Line } from './core/shapes/line';
 export { Point } from './core/shapes/point';
 export { Rectangle } from './core/shapes/rectangle';